daily-check/requirements.txt
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.4
fear-and-greed==0.4
frozendict==2.4.6
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-secret-manager==2.24.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multitasking==0.0.11
numpy==2.3.1
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.4
requests-cache==1.2.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.64
